+++ scripts/constants.js
@@ -3,7 +3,8 @@ export const DEFAULT_BATTLE_DAMAGE_DIRECTORY = "BattleDamage";
 
 export const FLAGS = {
   CONFIG: "config",
-  ORIGINAL_TEXTURE_SRC: "originalTextureSrc"
+  ORIGINAL_TEXTURE_SRC: "originalTextureSrc",
+  ORIGINAL_ACTOR_IMG: "originalActorImg"
 };
 
 export const SETTINGS = {
@@ -14,6 +15,7 @@ export const SETTINGS = {
   ONLY_GM_UPDATES: "onlyGmUpdates",
   INCLUDE_DEFEATED_IMAGE: "includeDefeatedImage",
   DEFAULT_IMAGE_DIRECTORY: "defaultImageDirectory",
+  UPDATE_ACTOR_PORTRAIT_FOR_TRACKERS: "updateActorPortraitForTrackers",
   REFRESH_COMBAT_TRACKER: "refreshCombatTracker",
   DEBUG_LOGGING: "debugLogging"
 };